refactor: split completions route into cohesive modules (wave 4a)
Break the completions god object (2313 lines, ~10 responsibilities) into
cohesive modules under routes/completions/ — pure move+import refactor,
zero behavior changes, public export surface of @polpo-ai/server unchanged
(completionRoutes, resumeProjectLoopRun, CompletionRouteDeps).

completions.ts: 2313 → 333 lines (auth, mode resolution, session
persistence, dispatch + CompletionRouteDeps only).

New modules:
- completions/schemas.ts (161) — Zod schemas + OpenAPI route definition
- completions/message-mapping.ts (166) — OpenAI ⇄ AI SDK message conversion
- completions/sse.ts (117) — SSE chunks, error envelopes, response shapes
- completions/tool-mapping.ts (171) — Polpo → AI SDK tool mapping, vault
  redaction, provider tool bookkeeping, client-side tool catalog
- completions/agent-step-runner.ts (236) — runAgentStepCompletion,
  buildRuntimeAgentPrompt, buildSummarizeFn, addUsage
- completions/project-loop-runner.ts (535) — runProjectLoopCompletion,
  buildLoopResumeState, resumeProjectLoopRun + loop-mode HTTP wiring
- completions/chat-handler.ts (777) — streaming/non-streaming multi-turn
  chat tool loop

Dedup with the task loop engine: the loop step helpers duplicated across
the completions loop runtime and packages/node loop-engine.ts
(buildLoopStepAgent, loopContextPrompt, maybeParseJson, normalizeToolInput,
stringifyLoopContext) now live once in
packages/core/src/loop/step-helpers.ts and are exported from the
@polpo-ai/core barrel. loop-engine.ts: 609 → 550 lines (−59), completions
runtime imports the same helpers instead of its own copies.
